integration test: test the edge cases, including the file size too big, files counts to many, there is no course infromation in the image, or some images have, some images doesn't have
